Fix/hr unit test5

+++ src/main/java/com/app/smartdrive/api/dto/HR/response/EmployeesResponseDto.java
@@ -1,5 +1,6 @@
 package com.app.smartdrive.api.dto.HR.response;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import com.app.smartdrive.api.dto.user.response.UserDto;
@@ -21,7 +22,7 @@ public class EmployeesResponseDto {
     private EnumClassHR.emp_type empType;
     private EnumClassHR.status empStatus;
     private EnumClassHR.emp_graduate empGraduate;
-    private Double empNetSalary;
+    private BigDecimal empNetSalary;
     private String empAccountNumber;
     private LocalDateTime empModifiedDate;
     private EmployeesJobTypeResponseDto jobType;
